#1890: extract coordinator/snapshot_refresh.rs (runtime-snapshot apply legs)
Second pure code-motion extraction for the #1890 split: move
refresh_fabric_links and the refresh_runtime_snapshot family (armed
leg, disarmed leg, and the shared _inner body — preflight policy
validation, manager-neighbor key rotation, #1873 tunnel-remap purge,
forwarding swap + worker-visible stores, WG/GRE reconcile-or-stop, CoS
owner-map refresh, neighbor warm pass) into
coordinator/snapshot_refresh.rs.

Pure motion with ZERO visibility changes: the public entry points keep
their `pub fn` signatures, refresh_runtime_snapshot_inner stays private
(both callers moved with it), and everything the body calls is either
already pub(crate)/pub(super) (purge_remapped_tunnel_sessions,
spawn_wg_control_threads, stop_all_*), private-in-coordinator and
therefore visible to this child module (queue_warm_pass,
refresh_cos_owner_worker_map_from_identities), or a mod.rs free helper
(tunnel_remap_purge_ids, log_wg_endpoint_set_transition) that stays in
mod.rs because reconcile/snapshot.rs and out-of-module tests reference
it at coordinator scope.

Validation: cargo build --release clean; warning set byte-identical to
the origin/master baseline. Full test gates run at the end of the
series.
